tests/wpt/web-platform-tests/largest-contentful-paint/video-poster.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: observe video poster image</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<script>
async_test(function (t) {
    assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
  const beforeLoad = performance.now();
  const observer = new PerformanceObserver(
    t.step_func_done(function(entryList) {
      assert_equals(entryList.getEntries().length, 1);
      const entry = entryList.getEntries()[0];
      const url = window.location.origin + '/images/blue.png';
      // blue.png is 133 by 106.
      const size = 133 * 106;
      checkImage(entry, url, 'the_poster', size, beforeLoad);
    })
  );
  observer.observe({type: 'largest-contentful-paint', buffered: true});
}, "Able to observe a video's poster image.");
</script>
<video id='the_poster' src='/media/test.mp4' poster='/images/blue.png'/>
